refactor id mapping
we have a bunch of different ids acrross aoi datasets. And there are certain expectation from the internal APIs we use regarding which ids they expect for which kind of aoi.

Primary goal of this commit is to makes sure whatever id field is expected by internal apis is also used across zeno by maintaining a common map of aoi sources to id column name.